+++ ruuvitag_sensor/common.py
@@ -1,9 +1,0 @@
-class RunFlag(object):
-    """
-    Wrapper for boolean run flag
-
-    Attributes:
-        running (bool): Defines if function should continue execution
-    """
-
-    running = True