+++ ruuvitag_sensor/ruuvi_rx.py
@@ -5,8 +5,7 @@ import time
 from concurrent.futures import ProcessPoolExecutor
 from rx.subjects import Subject
 
-from ruuvitag_sensor.ruuvi import RuuviTagSensor
-from ruuvitag_sensor.common import RunFlag
+from ruuvitag_sensor.ruuvi import RuuviTagSensor, RunFlag
 
 
 class RuuviTagReactive(object):